+++ Tests/NetworkLoggerPluginSpec.swift
@@ -1,7 +1,6 @@
 import Quick
 import Nimble
 import Moya
-import Result
 import enum Alamofire.AFError
 
 final class NetworkLoggerPluginSpec: QuickSpec {
@@ -105,7 +104,7 @@ final class NetworkLoggerPluginSpec: QuickSpec {
         }
 
         it("outputs an empty response message") {
-            let emptyResponseError = AFError.responseSerializationFailed(reason: .inputDataNil)
+            let emptyResponseError = AFError.responseSerializationFailed(reason: .inputFileNil)
             let result: Result<Moya.Response, MoyaError> = .failure(.underlying(emptyResponseError, nil))
 
             plugin.didReceive(result, target: GitHub.zen)
@@ -134,11 +133,11 @@ private class TestStreamRequest: RequestType {
         return request
     }
 
-    func authenticate(user: String, password: String, persistence: URLCredential.Persistence) -> Self {
+    func authenticate(username user: String, password: String, persistence: URLCredential.Persistence) -> Self {
         return self
     }
 
-    func authenticate(usingCredential credential: URLCredential) -> Self {
+    func authenticate(with credential: URLCredential) -> Self {
         return self
     }
 }
@@ -152,11 +151,11 @@ private class TestBodyRequest: RequestType {
         return request
     }
 
-    func authenticate(user: String, password: String, persistence: URLCredential.Persistence) -> Self {
+    func authenticate(username user: String, password: String, persistence: URLCredential.Persistence) -> Self {
         return self
     }
 
-    func authenticate(usingCredential credential: URLCredential) -> Self {
+    func authenticate(with credential: URLCredential) -> Self {
         return self
     }
 }
@@ -170,11 +169,11 @@ private class TestCurlBodyRequest: RequestType, CustomDebugStringConvertible {
         return request
     }
 
-    func authenticate(user: String, password: String, persistence: URLCredential.Persistence) -> Self {
+    func authenticate(username user: String, password: String, persistence: URLCredential.Persistence) -> Self {
         return self
     }
 
-    func authenticate(usingCredential credential: URLCredential) -> Self {
+    func authenticate(with credential: URLCredential) -> Self {
         return self
     }
 
@@ -188,11 +187,11 @@ private class TestNilRequest: RequestType {
         return nil
     }
 
-    func authenticate(user: String, password: String, persistence: URLCredential.Persistence) -> Self {
+    func authenticate(username user: String, password: String, persistence: URLCredential.Persistence) -> Self {
         return self
     }
 
-    func authenticate(usingCredential credential: URLCredential) -> Self {
+    func authenticate(with credential: URLCredential) -> Self {
         return self
     }
 }